rename kibana -> opensearch dashboards
