+++ .env-sample
@@ -1,5 +1,5 @@
 # you probably shouldn't change these
-FLASK_APP="kibana_cf_auth_proxy.app:create_app()"
+FLASK_APP="opensearch_dashboards_cf_auth_proxy.app:create_app()"
 FLASK_ENV=local
 
 # you might change these
