+++ DEVELOPMENT.md
@@ -1,3 +1,5 @@
+# Development
+
 ## Getting started
 
 Most development-related tasks are exposed in the `dev` script. To just get started developing,
