+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn --access-logfile - --error-logfile - --log-level info --timeout 300 "kibana_cf_auth_proxy.app:create_app()"
+web: gunicorn --access-logfile - --error-logfile - --log-level info --timeout 300 "opensearch_dashboards_cf_auth_proxy.app:create_app()"
