+++ dev
@@ -35,7 +35,7 @@ usage() {
     # upgrade version for specific requirement
     ${me} upgrade-requirements <requirement>
 
-    # run flask locally, (Bring Your Own [UAA, CF, Kibana])
+    # run flask locally, (Bring Your Own [UAA, CF, opensearch_dashboards])
     ${me} serve
 
     # run a local Opensearch + Opensearch Dashboards cluster in docker-compose
@@ -156,7 +156,7 @@ main() {
       ;;
     serve)
       set -o allexport; source "${dir}/.env"; set +o allexport
-      export FLASK_APP="kibana_cf_auth_proxy.app:create_app()"
+      export FLASK_APP="opensearch_dashboards_cf_auth_proxy.app:create_app()"
       
       ${python} -m flask run -p "${PORT}"
       ;;
