+++ docker/opensearch_dashboards/opensearch_dashboards.yml
@@ -8,7 +8,7 @@ opensearch.password: admin
 
 opensearch_security.multitenancy.enabled: true
 opensearch_security.multitenancy.tenants.preferred: ["Private", "Global"]
-opensearch_security.readonly_mode.roles: ["kibana_read_only"]
+opensearch_security.readonly_mode.roles: ["opensearch_dashboards_read_only"]
 
 # Use this setting if you are running opensearch dashboards without https
 opensearch_security.cookie.secure: false
