--- kibana_cf_auth_proxy/extensions.py
@@ -1,3 +1,0 @@
-from kibana_cf_auth_proxy.config import config_from_env
-
-config = config_from_env()
