+++ opensearch_dashboards_cf_auth_proxy/extensions.py
@@ -1,0 +1,3 @@
+from opensearch_dashboards_cf_auth_proxy.config import config_from_env
+
+config = config_from_env()
