+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = kibana_cf_auth_proxy
+name = opensearch_dashboards_cf_auth_proxy
 version = attr: src.VERSION
 description = Authentication proxy for use with Opensearch and Cloudfoundry
 long_description = file: README.md, CHANGELOG.md, LICENSE.md
