+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from kibana_cf_auth_proxy.app import create_app
+from opensearch_dashboards_cf_auth_proxy.app import create_app
 
 
 def pytest_configure(config):
