+++ tests/unit/test_config.py
@@ -1,5 +1,5 @@
 import pytest
-from kibana_cf_auth_proxy.config import config_from_env
+from opensearch_dashboards_cf_auth_proxy.config import config_from_env
 
 
 def test_config_loads(monkeypatch):
