+++ tests/unit/test_headers.py
@@ -1,4 +1,4 @@
-from kibana_cf_auth_proxy.headers import list_to_ext_header
+from opensearch_dashboards_cf_auth_proxy.headers import list_to_ext_header
 
 
 def test_spaces_to_header():
